@@ -1,6 +1,6 @@
 {
   "name": "@discordjs/builders",
-  "version": "1.5.0",
+  "version": "1.6.1",
   "description": "A set of builders that you can use when creating your bot",
   "scripts": {
     "test": "vitest run",
@@ -55,29 +55,29 @@
   },
   "homepage": "https://discord.js.org",
   "dependencies": {
-    "@discordjs/formatters": "^0.2.0",
+    "@discordjs/formatters": "^0.3.0",
     "@discordjs/util": "^0.2.0",
     "@sapphire/shapeshift": "^3.8.1",
-    "discord-api-types": "^0.37.35",
+    "discord-api-types": "^0.37.37",
     "fast-deep-equal": "^3.1.3",
     "ts-mixer": "^6.0.3",
     "tslib": "^2.5.0"
   },
   "devDependencies": {
-    "@favware/cliff-jumper": "^1.10.0",
+    "@favware/cliff-jumper": "^2.0.0",
     "@microsoft/api-extractor": "^7.34.4",
-    "@types/node": "16.18.13",
-    "@vitest/coverage-c8": "^0.29.1",
+    "@types/node": "16.18.23",
+    "@vitest/coverage-c8": "^0.29.8",
     "cross-env": "^7.0.3",
     "downlevel-dts": "^0.11.0",
-    "esbuild-plugin-version-injector": "^1.0.3",
-    "eslint": "^8.35.0",
-    "eslint-config-neon": "^0.1.40",
-    "eslint-formatter-pretty": "^4.1.0",
-    "prettier": "^2.8.4",
-    "tsup": "^6.6.3",
-    "typescript": "^4.9.5",
-    "vitest": "^0.29.1"
+    "esbuild-plugin-version-injector": "^1.1.0",
+    "eslint": "^8.37.0",
+    "eslint-config-neon": "^0.1.41",
+    "eslint-formatter-pretty": "^5.0.0",
+    "prettier": "^2.8.7",
+    "tsup": "^6.7.0",
+    "typescript": "^5.0.3",
+    "vitest": "^0.29.8"
   },
   "engines": {
     "node": ">=16.9.0"
