@@ -1,6 +1,6 @@
 {
   "name": "@discordjs/rest",
-  "version": "1.6.0",
+  "version": "1.7.0",
   "description": "The REST API for discord.js",
   "scripts": {
     "test": "vitest run",
@@ -53,29 +53,29 @@
   },
   "homepage": "https://discord.js.org",
   "dependencies": {
-    "@discordjs/collection": "^1.4.0",
+    "@discordjs/collection": "^1.5.0",
     "@discordjs/util": "^0.2.0",
     "@sapphire/async-queue": "^1.5.0",
     "@sapphire/snowflake": "^3.4.0",
-    "discord-api-types": "^0.37.35",
+    "discord-api-types": "^0.37.37",
     "file-type": "^18.2.1",
     "tslib": "^2.5.0",
-    "undici": "^5.20.0"
+    "undici": "^5.21.0"
   },
   "devDependencies": {
-    "@favware/cliff-jumper": "^1.10.0",
+    "@favware/cliff-jumper": "^2.0.0",
     "@microsoft/api-extractor": "^7.34.4",
-    "@types/node": "16.18.13",
-    "@vitest/coverage-c8": "^0.29.1",
+    "@types/node": "16.18.23",
+    "@vitest/coverage-c8": "^0.29.8",
     "cross-env": "^7.0.3",
-    "esbuild-plugin-version-injector": "^1.0.3",
-    "eslint": "^8.35.0",
-    "eslint-config-neon": "^0.1.40",
-    "eslint-formatter-pretty": "^4.1.0",
-    "prettier": "^2.8.4",
-    "tsup": "^6.6.3",
-    "typescript": "^4.9.5",
-    "vitest": "^0.29.1"
+    "esbuild-plugin-version-injector": "^1.1.0",
+    "eslint": "^8.37.0",
+    "eslint-config-neon": "^0.1.41",
+    "eslint-formatter-pretty": "^5.0.0",
+    "prettier": "^2.8.7",
+    "tsup": "^6.7.0",
+    "typescript": "^5.0.3",
+    "vitest": "^0.29.8"
   },
   "engines": {
     "node": ">=16.9.0"
