@@ -1,6 +1,6 @@
 {
   "name": "@sapphire/shapeshift",
-  "version": "3.8.1",
+  "version": "3.8.2",
   "description": "Blazing fast input validation and transformation ⚡",
   "author": "@sapphire",
   "license": "MIT",
@@ -21,8 +21,6 @@
     "format": "prettier --write \"{src,tests}/**/*.ts\"",
     "docs": "typedoc-json-parser",
     "test": "vitest run",
-    "test:watch": "vitest",
-    "update": "yarn upgrade-interactive",
     "build": "tsup",
     "clean": "node scripts/clean.mjs",
     "typecheck": "tsc -p tsconfig.eslint.json",
@@ -32,38 +30,42 @@
     "prepack": "yarn build && pinst --disable",
     "postpack": "pinst --enable"
   },
+  "dependencies": {
+    "fast-deep-equal": "^3.1.3",
+    "lodash": "^4.17.21"
+  },
   "devDependencies": {
-    "@commitlint/cli": "^17.3.0",
-    "@commitlint/config-conventional": "^17.3.0",
-    "@favware/cliff-jumper": "^1.9.0",
+    "@commitlint/cli": "^17.5.1",
+    "@commitlint/config-conventional": "^17.4.4",
+    "@favware/cliff-jumper": "^2.0.0",
     "@favware/npm-deprecate": "^1.0.7",
     "@sapphire/eslint-config": "^4.3.8",
-    "@sapphire/prettier-config": "^1.4.4",
+    "@sapphire/prettier-config": "^1.4.5",
     "@sapphire/ts-config": "^3.3.4",
-    "@types/jsdom": "^20.0.1",
-    "@types/lodash": "^4.14.191",
-    "@types/node": "^18.11.13",
-    "@typescript-eslint/eslint-plugin": "^5.46.0",
-    "@typescript-eslint/parser": "^5.46.0",
-    "@vitest/coverage-c8": "^0.25.7",
+    "@types/jsdom": "^21.1.1",
+    "@types/lodash": "^4.14.192",
+    "@types/node": "^18.15.11",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
+    "@vitest/coverage-c8": "^0.29.8",
     "cz-conventional-changelog": "^3.3.0",
-    "esbuild-plugins-node-modules-polyfill": "^1.0.7",
-    "eslint": "^8.29.0",
-    "eslint-config-prettier": "^8.5.0",
+    "esbuild-plugins-node-modules-polyfill": "^1.0.10",
+    "eslint": "^8.37.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "husky": "^8.0.2",
-    "jsdom": "^20.0.3",
-    "lint-staged": "^13.1.0",
+    "husky": "^8.0.3",
+    "jsdom": "^21.1.1",
+    "lint-staged": "^13.2.0",
     "pinst": "^3.0.0",
-    "prettier": "^2.8.1",
+    "prettier": "^2.8.7",
     "pretty-quick": "^3.1.3",
     "ts-node": "^10.9.1",
-    "tsup": "^6.5.0",
-    "typedoc": "^0.23.22",
-    "typedoc-json-parser": "^7.0.2",
-    "typescript": "^4.9.4",
-    "vite": "^4.0.0",
-    "vitest": "^0.25.7"
+    "tsup": "^6.7.0",
+    "typedoc": "^0.23.28",
+    "typedoc-json-parser": "^7.2.0",
+    "typescript": "^5.0.3",
+    "vite": "^4.2.1",
+    "vitest": "^0.29.8"
   },
   "repository": {
     "type": "git",
@@ -117,11 +119,7 @@
   },
   "resolutions": {
     "ansi-regex": "^5.0.1",
-    "minimist": "^1.2.7"
+    "minimist": "^1.2.8"
   },
-  "packageManager": "yarn@3.3.0",
-  "dependencies": {
-    "fast-deep-equal": "^3.1.3",
-    "lodash": "^4.17.21"
-  }
+  "packageManager": "yarn@3.5.0"
 }
