@@ -1,6 +1,6 @@
 {
   "name": "discord.js",
-  "version": "14.8.0",
+  "version": "14.9.0",
   "description": "A powerful library for interacting with the Discord API",
   "scripts": {
     "test": "yarn docs:test && yarn test:typescript",
@@ -49,32 +49,32 @@
   },
   "homepage": "https://discord.js.org",
   "dependencies": {
-    "@discordjs/builders": "^1.5.0",
-    "@discordjs/collection": "^1.4.0",
-    "@discordjs/formatters": "^0.2.0",
-    "@discordjs/rest": "^1.6.0",
+    "@discordjs/builders": "^1.6.0",
+    "@discordjs/collection": "^1.5.0",
+    "@discordjs/formatters": "^0.3.0",
+    "@discordjs/rest": "^1.7.0",
     "@discordjs/util": "^0.2.0",
     "@sapphire/snowflake": "^3.4.0",
     "@types/ws": "^8.5.4",
-    "discord-api-types": "^0.37.35",
+    "discord-api-types": "^0.37.37",
     "fast-deep-equal": "^3.1.3",
     "lodash.snakecase": "^4.1.1",
     "tslib": "^2.5.0",
-    "undici": "^5.20.0",
-    "ws": "^8.12.1"
+    "undici": "^5.21.0",
+    "ws": "^8.13.0"
   },
   "devDependencies": {
     "@discordjs/docgen": "^0.12.1",
-    "@favware/cliff-jumper": "^1.10.0",
-    "@types/node": "16.18.13",
+    "@favware/cliff-jumper": "^2.0.0",
+    "@types/node": "16.18.23",
     "dtslint": "^4.2.1",
-    "eslint": "^8.35.0",
-    "eslint-formatter-pretty": "^4.1.0",
-    "jest": "^29.4.3",
-    "prettier": "^2.8.4",
-    "tsd": "^0.25.0",
+    "eslint": "^8.37.0",
+    "eslint-formatter-pretty": "^5.0.0",
+    "jest": "^29.5.0",
+    "prettier": "^2.8.7",
+    "tsd": "^0.28.1",
     "tslint": "^6.1.3",
-    "typescript": "^4.9.5"
+    "typescript": "^5.0.3"
   },
   "engines": {
     "node": ">=16.9.0"
