@@ -1,12 +1,12 @@
-const { SlashCommandBuilder } = require('discord.js');
-const ping = require('./ping');
+import { SlashCommandBuilder } from 'discord.js';
+import { execute } from './ping';
 
 describe('Ping Command', () => {
     test('Reply with "Pong"', async () => {
         const mockInteraction = {
             reply: jest.fn(),
         }
-        await ping.execute(mockInteraction);
+        await execute(mockInteraction);
         expect(mockInteraction.reply).toHaveBeenCalledWith('Pong');
     })
 })
